+++ CHANGES.md
@@ -2,6 +2,7 @@
 - [FEATURE] Add StructuredDataUtility
 - [FEATURE] Add Recipe structured data for Detail view
 - [FEATURE] Extend Ingredient to link to a recipe
+- [FEATURE] Toggle WakeLock on CookingMode
 - [CHANGE] Recipe in List is an itemprop Article
 - [CHANGE] Use FlexForm/TypoScript media size settings